+++ sdk/textanalytics/Azure.AI.TextAnalytics/src/SentimentConfidenceScore.cs
@@ -6,9 +6,9 @@ namespace Azure.AI.TextAnalytics
     /// <summary>
     /// The sentiment confidence scores, by sentiment label.
     /// </summary>
-    public class SentimentConfidenceScorePerLabel
+    public class SentimentConfidenceScore
     {
-        internal SentimentConfidenceScorePerLabel(double positive, double neutral, double negative)
+        internal SentimentConfidenceScore(double positive, double neutral, double negative)
         {
             Positive = positive;
             Neutral = neutral;